Move UIFollowCamera to widgets/
