+++ include/fullscore/app_controller.h
@@ -8,9 +8,9 @@
 #include <fullscore/actions/queue_action.h>
 #include <fullscore/models/reference_cursor.h>
 #include <fullscore/widgets/command_bar.h>
+#include <fullscore/widgets/follow_camera.h>
 #include <fullscore/widgets/gui_score_editor.h>
 #include <fullscore/widgets/measure_inspector.h>
-#include <fullscore/follow_camera.h>
 #include <fullscore/playback_device_generic.h>
