+++ include/fullscore/widgets/follow_camera.h
