+++ src/actions/set_camera_target_action.cpp
@@ -4,7 +4,7 @@
 
 #include <fullscore/actions/set_camera_target_action.h>
 
-#include <fullscore/follow_camera.h>
+#include <fullscore/widgets/follow_camera.h>
